requirements.txt
asgiref==3.8.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
cssbeautifier==1.15.1
Django==5.1.4
django-appconf==1.0.6
django-compressor==4.5.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
djlint==1.36.3
EditorConfig==0.17.0
idna==3.10
jsbeautifier==1.15.1
json5==0.10.0
mysql-connector-python==9.1.0
pathspec==0.12.1
psycopg2-binary==2.9.10
pycparser==2.22
pyodbc==5.2.0
pyspnego==0.11.2
python-decouple==3.8
PyYAML==6.0.2
rcssmin==1.1.2
regex==2024.11.6
requests==2.32.3
requests_ntlm==1.3.0
rjsmin==1.2.2
six==1.17.0
sqlparse==0.5.3
sspilib==0.2.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0
